requirements.txt
streamlit
instaloader
pandas
googlesearch-python